Moves x86 code emitter into a separate file.
